Add IndexTuple IndexAttributeBitMapData array handling
This was previously omitted in error; the space used by bitmap arrays in
B-Tree index tuples (index tuples with 1 or more NULL attributes) was
incorrectly counted as a part of the tuple contents.

This oversight seems important, as adding this new field nicely
illustrates the potentially significant overhead of NULL attributes in
B-Tree indexes (alignment will tend to add rather a lot).