Add -x LSN threshold option
This option allows the user to specify an LSN that pg_hexedit uses to
determine if a given page is interesting.  If a page was modified on or
after that point (as determined by the page header LSN value), then
include it.  Otherwise, exclude it.

This is another option intended to provide a way to reduce the overhead
of having many tags in memory within wxHexEditor.  It isn't used by any
of the presupplied convenience scripts, though that may change in the
future.