Try to limp along in the event of an error
Return to the pg_filedump model of holding on in the event of an error.
Now that pg_hexedit uses stderr for errors, we can do better when there
is some isolated problem.

The convenience scripts will surface problems without launching
wxHexEditor. The user can make a decision to continue with opening the
relation manually, They will get whatever tags pg_hexedit managed to
produce.