Fix logic for posting lists/compression
Previously, we didn't check if a GIN IndexTuple was from a leaf page or
internal page before checking if it was compressed; the
GinItupIsCompressed() macro is only safe for leaf pages.  This
accidentally failed to fail because it was based on testing the block
number high bit, which was unlikely to be set in an internal page's
IndexTuple with typical pg_hexedit usage.  In other words, faulty code
mistook GinItupIsCompressed() as a check for whether or not the
containing page was a leaf page.

While at it, clarify our position on non-compressed posting lists within
leaf page IndexTuples, which can come from older Postgres versions.  We
still want to avoid representing that the posting lists are compressed
within leaf pages when that's wrong.  To keep things simple, we treat
any uncompressed posting list as tuple contents.