Add "verify block checksums when non-zero" flag
When the checksum field is non-zero, that must be because checksums were
enabled.  We check them at that point if the new -z option is in use.
We preserve the pg_filedump -k flag, which always checks checksums, even
for blocks with a zero checksum value.

This is now used by all convenience scripts, as an extra sanity check
that won't break workflows where checksums were not enabled at initdb
time. The user still gets most of the benefit of checksums when they
were enabled at initdb time, even when they didn't bother to modify a
convenience script by hand to specify -k.