Raise an error in event of unexpected special section change.
Postgres expects that the type of all blocks within a relfile segment
file match, since a relfile exists to store page images from one single
relation only.  I encountered a case in the wild where this expectation
was not met due to corruption, even though each block within the file
was well-formed as far as pg_hexedit could determine (specifically, an
nbtree index page appeared in a database's pg_attribute heap relfile).
It seems like a good idea for pg_hexedit to raise an error when there is
an inconsistency like this.