Add pg_relfilenodemapdata tool.
pg_relfilenodemapdata is a standalone program that prints information
from a specified pg_filenode.map (catalog relation to relfilenode
mapping) binary file.  It can be used to determine which relfilenode
corresponds to nailed/shared catalogs like pg_attribute and pg_database
when a working connection to the database cannot be established.