Fix bugs in GIN and hash annotations.
GIN non-data (main tree) leaf pages abuse the IndexTuple t_tid fields,
primarily to support posting lists.  Fix a bug in how this was annotated
by making a clearer distinction between posting lists and compressed
index tuples.  Also, this special use of t_tid fields is now explicitly
indicated in emitted GIN leaf page tags.

Stop attempting to emit a representation of tuples within a deleted GIN
page.  VACUUM seems to discard information about the basic type of the
page as part of page deletion, so this causes problems when we process a
posting tree deleted page as if it contained IndexTuples (as if it was a
pending list page).

Recognize hash bitmap pages as a distinct variety of ItemId-less page,
and add handling for that case.

These issues were noticed while adding a cross check of lp_len against
IndexTupleSize() to help with detection of torn pages.  That is also
introduced by this commit.