Give visual cues about xmin/xmax status.
Use special font colors to represent when xmin/xmax fields contain
special XIDs.  Represent this in xmin and xmax annotations, too.

Also make t_ctid field's font a non-contrasting color to deemphasize
t_ctid in the common case where t_ctid points to the tuple itself (where
tuple is the most recent tuple in an update chain).  This gives a visual
cue about whether or not a tuple is part of an update chain.