Use color to represent the details of line pointers.
Also add additional logic for xmax tags: make it apparent that the xmax
is for a locker xact, as opposed to an updater xact.

The overall effect of these changes is that it's now easier to follow
the structure of update chains on heap pages, especially when there is a
lot of bloat on the page.