Don't tag attributes that were truncated away.
nbtree pivot tuples in covering indexes do not contain non-key (INCLUDE)
attributes, and do not represent the absence of truncated attributes
using a NULL bitmap.  Avoid trying to emit tags for attributes that were
truncated away.  The number of attributes in pivot tuples is explicitly
represented where it doesn't match the number of key attributes for the
relation; we now check this.