Don't emit attribute tags when error encountered.
This seems to make pg_hexedit better able to cope with unsupported
access methods, such as zheap.  pg_hexedit ends up inferring that the
pages are of some kind of index access method, which is usually quite a
lot more useful than having no annotations at all.  The basic outline of
tuples will generally be represented approximately correctly for AMs
where the bufpage.c routines are used.

Clearly users can have no reasonable expectation of anything working in
this scenario, but it seems worthwhile to cope as best we can.