Print debug output when -x option used.
This is intended to make it easy to see how many blocks have not been
annotated when the -x option is used.  It should now be possible to
quickly iterate by tweaking the LSN passed to -x, until approximately
the desired number of blocks have been excluded from being tagged and
annotated.

Also print the return value for the pg_hexedit frontend utility in order
to enforce the idea that that matters when the convenience scripts are
run.  This also establishes the principle that debug or instrumentation
output from the pg_hexedit frontend program is normal and routine.